+++ tests/de/aidger/model/reports/ActivityEmploymentTest.java
@@ -94,7 +94,7 @@ public class ActivityEmploymentTest {
         contract.save();
 
         course = new Course();
-        course.setId(1);
+        course.setId((long) 1);
         course.setAdvisor("Tester");
         course.setDescription("Description");
         course.setFinancialCategoryId(fc.getId());
@@ -110,7 +110,7 @@ public class ActivityEmploymentTest {
         course.save();
 
         employment = new Employment();
-        employment.setId(1);
+        employment.setId((long) 1);
         employment.setAssistantId(assistant.getId());
         employment.setContractId(contract.getId());
         employment.setCourseId(course.getId());
