+++ tests/de/aidger/model/reports/ProtocolCreatorTest.java
@@ -116,7 +116,7 @@ public class ProtocolCreatorTest {
 
         activity = new Activity();
         activity.clearTable();
-        activity.setId(1);
+        activity.setId((long) 1);
         activity.setAssistantId(assistant.getId());
         activity.setContent("New assistant");
         activity.setCourseId(course.getId());
