+++ tests/de/aidger/utils/reports/ActivityReportHelperTest.java
@@ -103,7 +103,7 @@ public class ActivityReportHelperTest {
         contract.save();
 
         course = new Course();
-        course.setId(1);
+        course.setId((long) 1);
         course.setAdvisor("Tester");
         course.setDescription("Description");
         course.setFinancialCategoryId(fc.getId());
@@ -119,7 +119,7 @@ public class ActivityReportHelperTest {
         course.save();
 
         employment = new Employment();
-        employment.setId(1);
+        employment.setId((long) 1);
         employment.setAssistantId(assistant.getId());
         employment.setContractId(contract.getId());
         employment.setCourseId(course.getId());
